fix: Check for conversion or default values before registering
This solves an issue where a resolver that delegates to another type but
with a custom conversion may overwrite either its input or output
conversion. An example of such as case would be the OptionalResolver
that sometimes broke if both inputs and outputs of Optional were used in
the same schema.